Prefer to namespace assets under gnar
This commit prefers namespacing routes instead of scoping. This decision requires nesting the models, controllers, and views for `assets` under a `Gnar` modile. This is requires a broad code change up front but improves organization.
