--- app/controllers/assets_controller.rb
@@ -1,5 +1,0 @@
-class AssetsController < AuthenticatedController
-  def index
-    @assets = Asset.all.includes(:user)
-  end
-end
