+++ app/controllers/gnar/assets_controller.rb
@@ -1,0 +1,7 @@
+module Gnar
+  class AssetsController < AuthenticatedController
+    def index
+      @assets = Asset.all.includes(:user)
+    end
+  end
+end
