+++ app/views/gnar/assets/_asset.html.erb
