+++ app/views/gnar/assets/_table.html.erb
