+++ app/views/gnar/assets/index.html.erb
@@ -3,6 +3,6 @@
 <% end %>
 <div class="container flex flex-col h-screen w-full text-center">
   <div style="margin:1rem;">
-    <%= render "assets/table", assets: @assets %>
+    <%= render "gnar/assets/table", assets: @assets %>
   </div>
 </div>
