+++ spec/factories/assets.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :asset do
+  factory :asset, class: "Gnar::Asset" do
     approximate_purchase_date { Date.new(2022, 7, 21) }
     asset_type { "laptop" }
     mac_address { Faker::Internet.unique.mac_address }
